finora-backend/requirements.txt
django>=6.0.1
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.9.0
python-dotenv>=1.2.1
Pillow>=12.0.0
